Bump compose, refactor pane shared element semantics
