+++ gradle/libs.versions.toml
@@ -1,15 +1,15 @@
 [versions]
 androidGradlePlugin = "8.13.1"
-activity-compose = "1.12.1"
+activity-compose = "1.12.2"
 androidxAppCompat = "1.7.1"
 androidxBenchmark = "1.4.1"
 androidxCore = "1.17.0"
 androidxCollection = "1.5.0"
 dokka = "2.1.0"
 compose = "1.10.0"
-composeMultiplatform = "1.10.0-rc01"
+composeMultiplatform = "1.10.0-rc02"
 composeMultiplatformMaterial3 = "1.10.0-alpha05"
-lifecycleMultiplatform = "2.10.0-alpha06"
+lifecycleMultiplatform = "2.10.0-alpha07"
 junit4 = "4.13.2"
 kotlin = "2.2.21"
 kotlinxCoroutines = "1.10.2"
@@ -23,7 +23,7 @@ appcompatV7 = "28.0.0"
 googleMaterial = "1.13.0"
 materialIcons = "1.7.3"
 mavenPublish = "0.35.0"
-navigationEventComposeMultiplatform = "1.0.0-rc01"
+navigationEventComposeMultiplatform = "1.0.0-rc02"
 navigation3 = "1.0.0"
 navigation3Multiplatform = "1.0.0-alpha06"
 savedstateMultiplatform = "1.4.0"
