+++ library/compose-threepane/src/commonMain/kotlin/com/tunjid/treenav/compose/threepane/ThreePane.kt
@@ -16,15 +16,20 @@
 
 package com.tunjid.treenav.compose.threepane
 
+import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.EnterExitState
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.core.FiniteAnimationSpec
+import androidx.compose.animation.core.MutableTransitionState
+import androidx.compose.animation.core.Transition
+import androidx.compose.animation.core.rememberTransition
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import com.tunjid.treenav.Node
 import com.tunjid.treenav.compose.Adaptation
 import com.tunjid.treenav.compose.Adaptation.Swap
@@ -97,6 +102,41 @@ fun <Destination : Node> ContentTransform.adaptTo(
     paneScope: PaneScope<ThreePane, Destination>
 ): ContentTransform = if (paneScope.canAnimate()) this else NoContentTransform
 
+
+@Composable
+internal fun rememberStaticExitedAnimatedVisibilityScope(): AnimatedVisibilityScope {
+    val transition = rememberTransition(
+        remember {
+            MutableTransitionState(
+                initialState = EnterExitState.PostExit
+            )
+        }
+    )
+    return remember(transition) {
+        StaticAnimatedVisibilityScope(transition)
+    }
+}
+
+internal fun PaneScope<ThreePane, *>.canAnimateSecondary(): Boolean {
+    if (inPredictiveBack) return false
+    if (!paneState.adaptations.contains(PrimaryToSecondary)) return false
+    if (paneState.adaptations.contains(Adaptation.Pop)) return false
+
+    return true
+}
+
+private val PrimaryToSecondary = Swap(
+    from = ThreePane.Primary,
+    to = ThreePane.Secondary
+)
+
+
+private class StaticAnimatedVisibilityScope(
+    private val staticTransition: Transition<EnterExitState>
+) : AnimatedVisibilityScope {
+    override val transition: Transition<EnterExitState>
+        get() = staticTransition
+}
 
 private val RouteTransitionAnimationSpec: FiniteAnimationSpec<Float> = tween(
     durationMillis = 700
